+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.45.0](https://github.com/factorialco/f0/compare/f0-react-native-v0.44.0...f0-react-native-v0.45.0) (2026-03-25)
+
+
+### Features
+
+* **step:** add F0Step component with showcase and documentation ([#3757](https://github.com/factorialco/f0/issues/3757)) ([3a1f361](https://github.com/factorialco/f0/commit/3a1f3610e16afff4bde7079eb03589306377834b))
+
+## [0.44.0](https://github.com/factorialco/f0/compare/f0-react-native-v0.43.0...f0-react-native-v0.44.0) (2026-03-25)
+
+
+### Features
+
+* **chip:** introduce F0Chip component with showcase and styles ([#3755](https://github.com/factorialco/f0/issues/3755)) ([ad1738e](https://github.com/factorialco/f0/commit/ad1738e883df76f67d67bf4e24794a1f972e3c71))
+
+## [0.43.0](https://github.com/factorialco/f0/compare/f0-react-native-v0.42.0...f0-react-native-v0.43.0) (2026-03-25)
+
+
+### Features
+
+* **progress:** add F0Progress component ([#3742](https://github.com/factorialco/f0/issues/3742)) ([401edd5](https://github.com/factorialco/f0/commit/401edd5ed005ecd3ec20014cbcd5938d2c0a2a3b))
+
 ## [0.42.0](https://github.com/factorialco/f0/compare/f0-react-native-v0.41.0...f0-react-native-v0.42.0) (2026-03-18)